requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pyjwt
passlib[bcrypt]
python-dotenv
pip install pymysql python-dotenv
